scaffold src dir structure
